Merge Candidate

+++ src/mutations/organization/UpdateOrganizationMutation.js
@@ -7,7 +7,6 @@ import CreateContactInlineMutation from "../contact/CreateContactInlineMutation"
 import UpdateEmailMutation from "../UpdateEmailMutation";
 import UpdatePhoneMutation from "../UpdatePhoneMutation";
 import DeleteRelationshMutation from "../DeleteRelationshMutation";
-import RelationshipGroupContactQuery from "../RelationshipGroupContactQuery";
 
 const mutation = graphql`
     mutation UpdateOrganizationMutation($input: UpdateOrganizationInput!) {
@@ -86,14 +85,16 @@ export default function UpdateOrganizationMutation(organization, callback) {
                             if (contact.origin === "store") {
                                 DeleteRelationshMutation(contact.role_relation_id);
                             } else if (contact.origin === "new") {
-                                DeleteRelationshMutation(contact.role_obj.relation_id);
+                                if (contact.role_obj.relation_id) {
+                                    DeleteRelationshMutation(contact.role_obj.relation_id);
+                                }
                             }
                             UpdateContactInlineMutation(contact, organization.id, null, contact.role);
                             UpdateEmailMutation(contact.handle_id, contact.email, contact.email_obj);
                             UpdatePhoneMutation(contact.handle_id, contact.phone, contact.phone_obj);
                         }
                     } else if (contact.status === "remove") {
-                        RelationshipGroupContactQuery(organization.id, contact.handle_id, DeleteRelationshMutation);
+                        DeleteRelationshMutation(contact.role_relation_id);
                     }
                 });
                 callback.push("/community/organizations/" + organization.id);
